[DOP-34706] Moving fixtures and tests into separate files
